fix(tests): align config editor coverage with root CI
